# backend/wutonkGinBlog.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WutonkGinBlog",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
    "/user/login": {
      "post": {
        "summary": "用户登陆接口",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/loginReq"
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/loginResp"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/user/info": {
      "get": {
        "summary": "用户信息接口",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": false,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {}
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/infoResp"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    }
  },
  "components": {
    "schemas": {
      "loginReq": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "ID 编号"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "loginResp": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "required": [
          "token"
        ]
      },
      "infoResp": {
        "type": "object",
        "properties": {
          "nike_name": {
            "type": "string"
          }
        },
        "required": [
          "nike_name"
        ]
      }
    },
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}